Add new option to set logging level (#10)
The gpbackup_s3_plugin plugin does not set the logging level at all. This leads
to severe log pollution. Unfortunately, in order to pass the logging level with
which gpbackup was launched, we need to extend the plugin API. Therefore, this
patch simply adds a new option logger_verbosity to the plugin configuration
file, which allows us to set the desired logging level for the plugin.

The duplicate test for encryption has been replaced with a test for the new
option.